[ncl] update swift-ui screens with Host
