+++ apps/native-component-list/src/screens/UI/ColorPickerScreen.ios.tsx
@@ -1,6 +1,6 @@
-import { ColorPicker, Switch } from '@expo/ui/swift-ui';
+import { ColorPicker, Host, Switch, VStack } from '@expo/ui/swift-ui';
 import * as React from 'react';
-import { Dimensions, ScrollView, Text } from 'react-native';
+import { ScrollView, Text } from 'react-native';
 
 import { Page, Section } from '../../components/Page';
 
@@ -12,21 +12,21 @@ export default function ColorPickerScreen() {
       <Page>
         <Section title="Color Picker">
           <Text>Color: {color}</Text>
-          <Switch
-            label="Supports opacity"
-            value={supportsOpacity}
-            onValueChange={setSupportsOpacity}
-          />
-          <ColorPicker
-            label="Select a color"
-            selection={color}
-            supportsOpacity={supportsOpacity}
-            onValueChanged={setColor}
-            style={{
-              width: Dimensions.get('window').width - 20,
-              height: 100,
-            }}
-          />
+          <Host matchContents>
+            <VStack spacing={8}>
+              <Switch
+                label="Supports opacity"
+                value={supportsOpacity}
+                onValueChange={setSupportsOpacity}
+              />
+              <ColorPicker
+                label="Select a color"
+                selection={color}
+                supportsOpacity={supportsOpacity}
+                onValueChanged={setColor}
+              />
+            </VStack>
+          </Host>
         </Section>
       </Page>
     </ScrollView>
