+++ apps/native-component-list/src/screens/UI/ContextMenuScreen.ios.tsx
@@ -1,4 +1,4 @@
-import { Button, Picker, Switch, ContextMenu, Submenu } from '@expo/ui/swift-ui';
+import { Button, Host, Picker, Switch, ContextMenu, Submenu } from '@expo/ui/swift-ui';
 import { useVideoPlayer, VideoView } from 'expo-video';
 import * as React from 'react';
 import { View, StyleSheet, Text } from 'react-native';
@@ -22,116 +22,75 @@ export default function ContextMenuScreen() {
   return (
     <View>
       <Section title="Single-Press Context Menu" row>
-        <ContextMenu style={{ width: 150, height: 50 }}>
-          <ContextMenu.Items>
-            <Button
-              systemImage="person.crop.circle.badge.xmark"
-              onPress={() => console.log('Pressed1')}>
-              Hello
-            </Button>
-            <Button variant="bordered" systemImage="heart" onPress={() => console.log('Pressed2')}>
-              I love
-            </Button>
-            <Picker
-              label="Doggos"
-              options={['very', 'veery', 'veeery', 'much']}
-              variant="menu"
-              selectedIndex={selectedIndex}
-              onOptionSelected={({ nativeEvent: { index } }) => setSelectedIndex(index)}
-            />
-          </ContextMenu.Items>
-          <ContextMenu.Trigger>
-            <Button variant="bordered" style={{ width: 150, height: 50 }}>
-              Show Menu
-            </Button>
-          </ContextMenu.Trigger>
-        </ContextMenu>
-      </Section>
-      <Section title="Long-Press Context Menu" row>
-        <ContextMenu activationMethod="longPress" style={styles.longPressMenu}>
-          <ContextMenu.Items>
-            <Switch
-              value={switchChecked}
-              label="Do u love doggos?"
-              variant="checkbox"
-              onValueChange={setSwitchChecked}
-            />
-            <Switch
-              value={switch2Checked}
-              variant="switch"
-              label="Will u marry doggos?"
-              onValueChange={setSwitch2Checked}
-            />
-            <Button role="destructive" systemImage="hand.thumbsdown">
-              I don't like doggos 😡
-            </Button>
-            <Submenu button={<Button systemImage="heart.slash">Evil submenu</Button>}>
-              <Button>I hate</Button>
-              <Button>doggos</Button>
-              <Submenu button={<Button>👹Very evil submenu 👺</Button>}>
-                <Button>I KILL</Button>
-                <Button>DOGGOS</Button>
-              </Submenu>
-            </Submenu>
-          </ContextMenu.Items>
-          <ContextMenu.Trigger>
-            <View style={styles.longPressMenu}>
-              <VideoView player={player} style={styles.longPressMenu} contentFit="cover" />
-            </View>
-          </ContextMenu.Trigger>
-          <ContextMenu.Preview>
-            <View style={styles.preview}>
-              <Text>This is a preview</Text>
-            </View>
-          </ContextMenu.Preview>
-        </ContextMenu>
-      </Section>
-      {/* TODO: Bring back Android examples */}
-      {/* {Platform.OS === 'android' && (
-        <Section title="Colorful Context Menu">
-          <ContextMenu color="#e3b7ff">
-            <ContextMenu.Trigger>
-              <Button variant="bordered" style={{ width: 200, height: 50 }}>
-                Show Colorful Menu
-              </Button>
-            </ContextMenu.Trigger>
+        <Host style={{ width: 150, height: 50 }}>
+          <ContextMenu>
             <ContextMenu.Items>
-              <Button variant="bordered" color="#ff0000">
-                I'm red!
+              <Button
+                systemImage="person.crop.circle.badge.xmark"
+                onPress={() => console.log('Pressed1')}>
+                Hello
               </Button>
               <Button
                 variant="bordered"
-                elementColors={{ containerColor: '#0000ff', contentColor: '#00ff00' }}>
-                My text is green!
+                systemImage="heart"
+                onPress={() => console.log('Pressed2')}>
+                I love
               </Button>
+              <Picker
+                label="Doggos"
+                options={['very', 'veery', 'veeery', 'much']}
+                variant="menu"
+                selectedIndex={selectedIndex}
+                onOptionSelected={({ nativeEvent: { index } }) => setSelectedIndex(index)}
+              />
+            </ContextMenu.Items>
+            <ContextMenu.Trigger>
+              <Button variant="bordered">Show Menu</Button>
+            </ContextMenu.Trigger>
+          </ContextMenu>
+        </Host>
+      </Section>
+      <Section title="Long-Press Context Menu" row>
+        <Host style={styles.longPressMenu}>
+          <ContextMenu activationMethod="longPress">
+            <ContextMenu.Items>
               <Switch
                 value={switchChecked}
-                label="I'm very colorful!"
+                label="Do u love doggos?"
                 variant="checkbox"
-                elementColors={{
-                  checkedColor: '#ff0000',
-                  disabledCheckedColor: '#00ff00',
-                  uncheckedColor: '#0000ff',
-                  checkmarkColor: '#ffff00',
-                }}
                 onValueChange={setSwitchChecked}
               />
               <Switch
                 value={switch2Checked}
                 variant="switch"
-                label="Switches can be colorul too!"
+                label="Will u marry doggos?"
                 onValueChange={setSwitch2Checked}
-                elementColors={{
-                  checkedThumbColor: '#ff0000',
-                  checkedTrackColor: '#00ff00',
-                  uncheckedThumbColor: '#0000ff',
-                  uncheckedTrackColor: '#ffff00',
-                }}
               />
+              <Button role="destructive" systemImage="hand.thumbsdown">
+                I don't like doggos 😡
+              </Button>
+              <Submenu button={<Button systemImage="heart.slash">Evil submenu</Button>}>
+                <Button>I hate</Button>
+                <Button>doggos</Button>
+                <Submenu button={<Button>👹Very evil submenu 👺</Button>}>
+                  <Button>I KILL</Button>
+                  <Button>DOGGOS</Button>
+                </Submenu>
+              </Submenu>
             </ContextMenu.Items>
+            <ContextMenu.Trigger>
+              <View style={styles.longPressMenu}>
+                <VideoView player={player} style={styles.longPressMenu} contentFit="cover" />
+              </View>
+            </ContextMenu.Trigger>
+            <ContextMenu.Preview>
+              <View style={styles.preview}>
+                <Text>This is a preview</Text>
+              </View>
+            </ContextMenu.Preview>
           </ContextMenu>
-        </Section>
-      )} */}
+        </Host>
+      </Section>
     </View>
   );
 }
