+++ apps/native-component-list/src/screens/UI/FormScreen.android.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { ScrollView, Text, View } from 'react-native';
 
-export default function SectionScreen() {
+export default function FormScreen() {
   return (
     <ScrollView style={{ flex: 1 }}>
       <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
@@ -11,6 +11,6 @@ export default function SectionScreen() {
   );
 }
 
-SectionScreen.navigationOptions = {
-  title: 'Section',
+FormScreen.navigationOptions = {
+  title: 'Form',
 };
