+++ apps/native-component-list/src/screens/UI/SliderScreen.ios.tsx
@@ -1,4 +1,4 @@
-import { Slider } from '@expo/ui/swift-ui';
+import { Host, Slider } from '@expo/ui/swift-ui';
 import * as React from 'react';
 
 import { Page, Section } from '../../components/Page';
@@ -8,35 +8,38 @@ export default function SliderScreen() {
   return (
     <Page>
       <Section title="Regular slider">
-        <Slider
-          style={{ minHeight: 60 }}
-          value={value}
-          onValueChange={(value) => {
-            setValue(value);
-          }}
-        />
+        <Host matchContents>
+          <Slider
+            value={value}
+            onValueChange={(value) => {
+              setValue(value);
+            }}
+          />
+        </Host>
       </Section>
       <Section title="Stepped slider">
-        <Slider
-          style={{ minHeight: 60 }}
-          value={value}
-          steps={5}
-          onValueChange={(value) => {
-            setValue(value);
-          }}
-        />
+        <Host matchContents>
+          <Slider
+            value={value}
+            steps={5}
+            onValueChange={(value) => {
+              setValue(value);
+            }}
+          />
+        </Host>
       </Section>
       <Section title="Colorful slider">
-        <Slider
-          style={{ minHeight: 60 }}
-          value={value}
-          min={-1}
-          max={5}
-          color="#ff0000"
-          onValueChange={(value) => {
-            setValue(value);
-          }}
-        />
+        <Host matchContents>
+          <Slider
+            value={value}
+            min={-1}
+            max={5}
+            color="#ff0000"
+            onValueChange={(value) => {
+              setValue(value);
+            }}
+          />
+        </Host>
       </Section>
     </Page>
   );
