+++ apps/native-component-list/src/screens/UI/TextInputScreen.ios.tsx
@@ -1,4 +1,4 @@
-import { TextField, TextFieldRef, SecureField, Button } from '@expo/ui/swift-ui';
+import { Button, Host, TextField, TextFieldRef, SecureField } from '@expo/ui/swift-ui';
 import * as React from 'react';
 import { Text, TextInput as RNTextInput } from 'react-native';
 
@@ -14,58 +14,70 @@ export default function TextInputScreen() {
       <Section title="Current value">
         <Text>{JSON.stringify(value)}</Text>
       </Section>
-      <Button
-        onPress={async () => {
-          textRef.current?.setText('Hello there!');
-          secureRef.current?.setText('123');
-        }}>
-        Set text
-      </Button>
+      <Host matchContents>
+        <Button
+          onPress={async () => {
+            textRef.current?.setText('Hello there!');
+            secureRef.current?.setText('123');
+          }}>
+          Set text
+        </Button>
+      </Host>
       <Section title="Text Input">
-        <TextField
-          ref={textRef}
-          autocorrection={false}
-          defaultValue="hey there"
-          onChangeText={setValue}
-        />
+        <Host matchContents>
+          <TextField
+            ref={textRef}
+            autocorrection={false}
+            defaultValue="hey there"
+            onChangeText={setValue}
+          />
+        </Host>
       </Section>
       <Section title="Multiline Text Input">
-        <TextField
-          multiline
-          numberOfLines={5}
-          autocorrection={false}
-          allowNewlines={false}
-          defaultValue="hey there"
-          onChangeText={setValue}
-        />
+        <Host matchContents>
+          <TextField
+            multiline
+            numberOfLines={5}
+            autocorrection={false}
+            allowNewlines={false}
+            defaultValue="hey there"
+            onChangeText={setValue}
+          />
+        </Host>
       </Section>
       <Section title="Phone Text Input">
-        <TextField
-          multiline
-          numberOfLines={5}
-          keyboardType="phone-pad"
-          autocorrection={false}
-          defaultValue="324342324"
-          onChangeText={setValue}
-        />
+        <Host matchContents>
+          <TextField
+            multiline
+            numberOfLines={5}
+            keyboardType="phone-pad"
+            autocorrection={false}
+            defaultValue="324342324"
+            onChangeText={setValue}
+          />
+        </Host>
       </Section>
       <Section title="Multiline, allowNewlines Text Input">
-        <TextField
-          multiline
-          numberOfLines={5}
-          allowNewlines
-          autocorrection={false}
-          defaultValue="hey there"
-          onChangeText={setValue}
-        />
+        <Host matchContents>
+          <TextField
+            multiline
+            numberOfLines={5}
+            allowNewlines
+            autocorrection={false}
+            defaultValue="hey there"
+            onChangeText={setValue}
+          />
+        </Host>
       </Section>
       <Section title="Secure Text Input">
-        <SecureField
-          ref={secureRef}
-          defaultValue="hey there"
-          onChangeText={setValue}
-          keyboardType="numeric"
-        />
+        <Host matchContents>
+          <SecureField
+            ref={secureRef}
+            defaultValue="hey there"
+            onChangeText={setValue}
+            keyboardType="numeric"
+          />
+        </Host>
       </Section>
       <Section title="RN Text Input">
         <RNTextInput multiline numberOfLines={5} value={value} onChangeText={setValue} />
