+++ apps/native-component-list/src/screens/UI/UIScreen.android.tsx
@@ -51,11 +51,11 @@ export const UIScreens = [
     },
   },
   {
-    name: 'Section component',
-    route: 'ui/section',
+    name: 'Form component',
+    route: 'ui/form',
     options: {},
     getComponent() {
-      return optionalRequire(() => require('./SectionScreen'));
+      return optionalRequire(() => require('./FormScreen'));
     },
   },
   {
@@ -132,7 +132,7 @@ export default function UIScreen() {
       route: `/components/${screen.route}`,
     };
   });
-  return <ComponentListScreen apis={apis} sort={false} />;
+  return <ComponentListScreen apis={apis} />;
 }
 
 UIScreen.navigationOptions = {
