+++ apps/native-component-list/src/screens/UI/UIScreen.ios.tsx
@@ -35,11 +35,11 @@ export const UIScreens = [
     },
   },
   {
-    name: 'Section component',
-    route: 'ui/section',
+    name: 'Form component',
+    route: 'ui/form',
     options: {},
     getComponent() {
-      return optionalRequire(() => require('./SectionScreen'));
+      return optionalRequire(() => require('./FormScreen'));
     },
   },
   {
@@ -115,11 +115,11 @@ export const UIScreens = [
     },
   },
   {
-    name: 'SwiftUI primitives',
-    route: 'ui/swiftui-primitives',
+    name: 'Hosting RN Views',
+    route: 'ui/hosting-rn-views',
     options: {},
     getComponent() {
-      return optionalRequire(() => require('./SwiftUIPrimitivesScreen'));
+      return optionalRequire(() => require('./HostingRNViewsScreen'));
     },
   },
   {
@@ -140,7 +140,7 @@ export default function UIScreen() {
       route: `/components/${screen.route}`,
     };
   });
-  return <ComponentListScreen apis={apis} sort={false} />;
+  return <ComponentListScreen apis={apis} sort />;
 }
 
 UIScreen.navigationOptions = {
