+++ apps/native-component-list/src/screens/UI/UIScreen.tsx
@@ -5,10 +5,7 @@ export const UIScreens = [];
 export default function UIScreen() {
   return (
     <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
-      <Text>
-        Expo UI is only available on Android and iOS currently.Example temporarily disabled on
-        Android.
-      </Text>
+      <Text>Expo UI is only available on Android and iOS currently.</Text>
     </View>
   );
 }
